Moved implicit to avoid collisions with CS repo code.
