+++ cpgqueryingtests/src/test/scala/io/shiftleft/cpgqueryingtests/steps/CDataFlowTests.scala
@@ -7,7 +7,7 @@ import io.shiftleft.passes.dataflows.steps.{FlowPrettyPrinter}
 import io.shiftleft.codepropertygraph.generated.{NodeKeys, nodes}
 import io.shiftleft.queryprimitives.steps.Implicits._
 import io.shiftleft.queryprimitives.steps._
-import io.shiftleft.passes.dataflows._
+import io.shiftleft.passes.dataflows.Implicits._
 
 import scala.language.implicitConversions
