+++ semanticcpg/src/main/scala/io/shiftleft/passes/dataflows/Implicits.scala
@@ -1,14 +1,13 @@
-package io.shiftleft.passes
+package io.shiftleft.passes.dataflows
 
-import io.shiftleft.passes.dataflows.steps.TrackingPoint
 import io.shiftleft.codepropertygraph.generated.nodes
+import io.shiftleft.passes.dataflows.steps.TrackingPoint
 import io.shiftleft.queryprimitives.steps.Steps
 import io.shiftleft.queryprimitives.steps.Implicits._
-import scala.language.implicitConversions
-
 import shapeless.HList
 
-package object dataflows {
+object Implicits {
+
   implicit def toTrackingPoint[X <% Steps[NodeType, Labels],
   NodeType <: nodes.TrackingPoint,
   Labels <: HList](steps: X): TrackingPoint[Labels] =
